rework for mock interface - rework the timer interface to behave similarly to the ruleset interface
