+++ src/pulsarity/interface/__init__.py
@@ -3,10 +3,9 @@ Hardware Interfaces
 """
 
 from pulsarity.interface.timer_interface import (
-    Action,
     BasicLapData,
     BasicSignalData,
     TimerSetting,
 )
 
-__all__ = ["Action", "BasicLapData", "BasicSignalData", "TimerSetting"]
+__all__ = ["BasicLapData", "BasicSignalData", "TimerSetting"]
