+++ tests/conftest.py
@@ -22,7 +22,7 @@ from pulsarity.database import (
     setup_default_objects,
 )
 from pulsarity.events.broker import EventBroker
-from pulsarity.interface.timer_manager import TimerInterfaceManager
+from pulsarity.interface.timer_interface import TimerInterfaceManager
 from pulsarity.race.manager import RaceManager
 from pulsarity.race.ruleset import RaceRulesetManager
 from pulsarity.utils import background
